Move each crossref class to it's own file and sort the attributes.
