+++ asseeibot/models/crossref/author.py
@@ -1,0 +1,10 @@
+from typing import Optional, List, Any
+
+from pydantic import BaseModel
+
+
+class CrossrefAuthor(BaseModel):
+    given: Optional[str]
+    family: Optional[str]
+    sequence: Optional[str]
+    affiliation: Optional[List[Any]]
